requirements.txt
numpy>=1.19.0
scipy>=1.5.0
scikit-image>=0.17.0
dask[distributed]>=2021.0.0
pyyaml>=5.4.0
rasterio>=1.2.0